Refactored blog form into separate component
